feat(migration): add AtaOwner column to TokenAccounts table
- Introduced a new migration to add the AtaOwner column to the TokenAccounts table.
- Updated the TokenAccounts enum to include AtaOwner.
- Registered the new migration in the standard migration module.

fix(monitor): optimize lag calculation and clean up code

- Simplified the lag calculation in the continously_monitor_photon function.
- Removed unnecessary code and improved readability in update_latest_slot and verify_queue_hash_chains functions.

refactor(monitor): enhance logging and improve error handling

- Added debug logging for parsed state trees in process_tree_account.
- Improved error handling in verify_queue_hash_chains and log_divergence functions.

feat(openapi): add new endpoints for account interfaces

- Implemented /getAccountInterfaces and /getAtaInterface endpoints in the OpenAPI specification.
- Updated the API response structures to accommodate new functionalities.

fix(snapshot): streamline error handling in snapshot loading

- Refactored error handling in snapshot loading and streaming functions for better clarity and consistency.

test(tests): enhance integration tests for token and mint data

- Updated integration tests to reflect changes in discriminator handling and added helper functions for padded values.
- Adjusted snapshot tests to include new tree type information for mint interfaces.

+++ src/ingester/startup_cleanup.rs
@@ -39,7 +39,7 @@ pub async fn cleanup_stale_address_queues(
             .order_by_desc(indexed_trees::Column::LeafIndex)
             .one(db)
             .await?
-            .map(|t| (t.leaf_index + 1) as i64)
+            .map(|t| t.leaf_index + 1)
             .unwrap_or(1);
 
         // Address queue indices are 0-based, tree indices are 1-based
